Extract workbench contract boundaries
